web/gridt/tests/integration/resources/test_user.py
import lorem
import json

from gridt.tests.base_test import BaseTest
from gridt.models.user import User
from gridt.resources.user import BioResource


class UserResourceTest(BaseTest):
    def test_update_bio(self):
        with self.app_context():
            new_bio = lorem.paragraph()

            movement = self.create_movement()
            user = self.create_user_in_movement(movement, generate_bio=True)

            resp = self.request_as_user(
                self.users[0], "PUT", "/bio", json={"bio": new_bio}
            )

            self.assertEqual(resp.get_json(), {"message": "Bio successfully changed."})
            self.assertEqual(user.bio, new_bio)

    def test_update_bio_improperly(self):
        with self.app_context():
            movement = self.create_movement()
            user = self.create_user_in_movement(movement, generate_bio=True)

            resp = self.request_as_user(self.users[0], "PUT", "/bio", json={})

            self.assertEqual(resp.get_json(), {"message": "Bad request."})
